refactor: optimize Service Topology panels for better screen utilization
- Remove redundant chart titles (panels already have titles)
- Fix ServiceTopologyGraph to use full available viewport space
- Remove fixed height constraints and use flexbox layout
- Eliminate extra Card wrapper from graph component
- Ensure consistent panel formatting across all three columns
